+++ src/constants.ts
@@ -1,8 +1,6 @@
+import { CommandType, SimplePromptPluginSettings } from "./types";
 
 // PROMPTS
-
-import { SimplePromptPluginSettings } from "./types";
-
 export const DEFAULT_CURSOR_PROMPT_TEMPLATE = `
 You are a helpful AI assistant that can, given a piece of text and a request generate an answer using markdown.
 Include headers, lists, checkboxes, and other markdown elements in your answer when it makes sense.
@@ -35,7 +33,7 @@ Request:
 ==================
 <QUERY>
 ==================
-Answer:`
+Answer:`;
 
 export const DEFAULT_SELECTION_PROMPT_TEMPLATE = `
 You are a helpful AI assistant that can, given a piece of text and a request generate an answer using markdown.
@@ -66,7 +64,7 @@ Request:
 ==================
 <REQUEST>
 ==================
-Answer:`
+Answer:`;
 
 export const DEFAULT_REWRITE_DOCUMENT_TEMPLATE = `
 You are a helpful AI assistant who is an expert in rewriting text. Given a markdown document and a request, you can generate a new version of the document.
@@ -98,14 +96,38 @@ Request:
 ==================
 <REQUEST>
 ==================
-Answer:`
+Answer:`;
 
-
-// COMMANDS
-export const SETTINGS_SET_API_KEY_COMMAND_NAME = "Set API key";
+// PROMPT COMMANDS
 export const SELECTION_COMMAND_NAME = "Rewrite selection";
 export const DOC_COMMAND_NAME = "Rewrite document";
 export const CURSOR_COMMAND_NAME = "Generate content at cursor";
+
+export const SELECTION_COMMAND_SUBTITLE = "Write your prompt to rewrite the selected text. Ex. 'Add more options', 'summarize the text', etc.";
+export const DOC_COMMAND_SUBTITLE = "Write you request to rewrite the entire document. Ex. 'Structure the document better', 'Add more examples', etc.";
+export const CURSOR_COMMAND_SUBTITLE = "Give request to generate content at the cursor. Ex. 'Write a joke', 'Explain the concept of ...', etc.";
+
+export const PROMPT_COMMANDS: {
+    name: string;
+    id: string;
+    type: CommandType;
+}[] = [
+    {
+        name: SELECTION_COMMAND_NAME,
+        id: "prompt-generate-content-from-selection",
+        type: "selection",
+    },
+    {
+        name: CURSOR_COMMAND_NAME,
+        id: "prompt-generate-content-at-cursor",
+        type: "cursor",
+    },
+    {
+        name: DOC_COMMAND_NAME,
+        id: "prompt-rewrite-document",
+        type: "document",
+    },
+];
 
 // SETTINGS
 export const DEFAULT_SETTINGS: SimplePromptPluginSettings = {
@@ -113,6 +135,7 @@ export const DEFAULT_SETTINGS: SimplePromptPluginSettings = {
     model: "gpt-3.5-turbo",
     recentPrompts: [],
     recentsLimit: 5,
+    recentPromptsEnabled: true,
     promptTemplates: {
         selection: DEFAULT_SELECTION_PROMPT_TEMPLATE,
         cursor: DEFAULT_CURSOR_PROMPT_TEMPLATE,
@@ -120,3 +143,10 @@ export const DEFAULT_SETTINGS: SimplePromptPluginSettings = {
     },
     streaming: false,
 };
+
+// SETTINGS COMMANDS
+export const SETTINGS_SET_API_KEY_COMMAND_NAME = "Set API key";
+export const SETTINGS_CHANGE_LLM_MODEL_COMMAND_NAME = "Change LLM";
+export const SETTINGS_TOGGLE_STREAMING_COMMAND_NAME = "Toggle streaming";
+export const SETTINGS_TOGGLE_RECENT_PROMPTS_COMMAND_NAME =
+    "Toggle Enable/Disable Recent Prompts";